Small refactoring for GameChoice and choices output on screen
